Workaround ExtUtils::Liblist stripping link library arguments.
As per https://github.com/Perl-Toolchain-Gang/ExtUtils-MakeMaker/issues/277 ,
EU:L does the wrong thing and assumes its missing, even when it is not.

On older GCC's this appears to have worked by magic, as it looks like
the "#include " directive anywhere in the source automatically activated
the link-library being injected by the compiler.

But since at least GCC 5.4, omitting an -lquadmath directive results
in a failed compiliation:

- https://rt.cpan.org/Ticket/Display.html?id=116520

And apparently, on some targets, the libraries -lquadmath and -lm may
not actually physically exist, but are faked into existence by the
compiler, and EU:L will always make the wrong choice, and this will
break compilation on those targets.

Hence, ensuring -lm is present is also done in this commit.